Use pytest argument parametrisation
This replaces several cases where we had a single, private method being
invoked by several test methods, which otherwise were the same, and a
few tests where we called a single private method several times with
different input values.

Even though we stopped using unittest, we still have many tests that are
part of Test* classes. As part of this commit, and where easy to do due
to the removal of duplicated methods that are now parametrised, we have
moved some tests to be free-standing functions instead of methods in a
test class.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>